Merges

+++ 1.5.06 with T-6638/xrGame/Bolt.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "bolt.h"
 #include "ParticlesObject.h"
-#include "PhysicsShell.h"
+#include "../xrphysics/PhysicsShell.h"
 #include "xr_level_controller.h"
 
 CBolt::CBolt(void) 
@@ -35,7 +35,7 @@ bool CBolt::Useful() const
 	return false;
 }
 
-bool CBolt::Action(s32 cmd, u32 flags) 
+bool CBolt::Action(u16 cmd, u32 flags) 
 {
 	if(inherited::Action(cmd, flags)) return true;
 /*
