+++ 1.5.06 with T-6638/xrGame/Bolt.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "missile.h"
-#include "DamageSource.h"
+#include "../xrphysics/DamageSource.h"
 class CBolt :
 					public CMissile,
 					public IDamageSource
@@ -17,7 +17,7 @@ public:
 	virtual	u16	 Initiator			();
 
 	virtual void Throw				();
-	virtual bool Action				(s32 cmd, u32 flags);
+	virtual bool Action				(u16 cmd, u32 flags);
 	virtual bool Useful				() const;
     virtual void activate_physic_shell	();
